requirements.txt
anthropic>=0.25.0
requests>=2.31.0
edge-tts>=6.1.9
Pillow>=10.0.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
